Workaround to ensure vue-cli knows the app is running inside of a container
The latest version of docker changed a file that the vue-cli service was relying on
to indicate whether or not the app is running inside of a container. Without that
indicator, hot module reload does not work b/c the dev server tries to connect to the
docker IP instead of localhost. Setting this env variable is another check that the
vue-cli service does to see if the app is running inside of a container.