Dependency update
Despite the relatively large jump in some of these dependencies, there
were very few breaking changes that we needed to resolve (most notably
the change of the `vue.config.js` options from `disableHostCheck` to
`allowedHosts`). I removed a few deps that it looked like we weren't
using anyway, and we had two

There was an interesting permissions issue that is occurring only on
Linux machines as far as we can tell. We handled it in a change to the
`Dockerfile` that makes the `node_modules/.cache` directory world
writable (not ideal solution) but would like to dig into this more in
the future, outside of this release cycle.