Use ratios of individual processed blocks/gaps to calculate backbone offsets for genes
The gapless block may have a different ratio than the blocks/gaps that make it up. To more
accurately determine where the backbone start and stop of a gene is, this commit makes a change
to find the processed synteny section where the gene starts and the processed synteny section where
the gene stops. Then it calculates the backbone start offset and backbone stop offset respective to
each of those section's backbone-to-synteny-section bp ratios.