Added working Search suggestions
The Search Gene process looks at the geneList in Main now through the
lens of the SelectedDataPanel.props, but selecting an off-backbone gene
from the list will not change the Detailed panel because it cannot find
the gene (and its orthologs, if present) on the SVGViewbox data
structures. This is intentional. Once we have all components looking at
the geneList (and blockList) in Main through their appropriate .props,
this will be resolved.