WIP: migrating setDetailedBasePairRange to request
I have not been able to isolate all of the navigation operations because
there appears to be some overlap with the "setBackboneSelection" and the
"setDetailedBasePairRange" work. My suspicion is that these two things
can be combined in the course of this implementation as they shouldn't
both be needed and are likely a hold over from the previous
implementation that had a separate concept of the selection and the
visable region.

The initial backbone load pulls in the backbone genes and that is the
only time it occurs (rightfully). But this appears to be triggered by
some actions of the Configuration view that I will need to track down.

From the work so far, it does appear the memory management is resolved,
and the performance is greatly improved.