Added block processing on initial load
As expected, this has drastically increased the load time for the Main
processing and it is now greater than the load time for the SVGViewbox
overview loading in a lot of situations. This call with genes and
orthologs always will be a slow process we cannot control in our code.
However, the structures created from this process are completed and done
before the detailed panel is ready in most cases. Once the processing
(and SVG element creation / DOM manipulation) is reduced because of
these new structures, we should see a major speed increase overall.