Initial re-implementation of BackboneSet
Using the data structures from Main, it was relatively easy to convert
the `updateDetailPanel()` method to remove API calls and rely only on
visible data from our props. Lots of code was touched in this
re-implementation because I tried to clean up everything I could while I
was working on it.