WIP: Selection of genes / labels
This code isn't currently handling orthologs because of the changes in
memory structures (and the fact that we aren't properly loading
orthologs into our new structures ;) )

Some desired changes to the code are noted as well. Shift handling isn't
quite right as it doesn't "toggle" a gene, just resets the entire
selection if a previously selected gene is shift-clicked. This behavior
is better than nothing though, but also missing from the labels. And the
code is duplicated in two different SVG components (Backbone and
Section). We will need to address that for maintainability.